Task/cambio de lugar archivo Slider y stylesSliderComponent
